requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.3.1
tweepy==4.14.0
urllib3==2.2.3
